backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==4.2.20
djangorestframework==3.16.0
face-recognition-models==0.3.0
numpy==2.0.2
pillow==11.1.0
psycopg==3.2.6
psycopg-binary==3.2.6
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.0
